feat: implement question bank and refactor backend/frontend architecture
1. Complete Question Bank Page implementation
   - Create, Import, Export, Edit, Copy, Delete question operations

2. Refactor backend main.py into modular routers
   - Split into admin.py, recruiter.py, candidate.py
   - Implement persistence for Projects and Positions (CRUD)
   - Fix utils.py data directory path resolution

3. Verify and integrate Projects/Positions operations
   - Connect frontend to backend CRUD endpoints
   - Add proper error handling and data refresh

4. Modularize API services in both portals
   - Split api.ts into domain-specific service files
   - Improve code organization and maintainability
   - Maintain backward compatibility

All builds passing, full verification completed.
