+++ Frontend/candidate-portal/src/services/candidate.service.ts
@@ -1,0 +1,7 @@
+import { fetchAPI } from './client';
+
+export const candidateService = {
+    getHome: async () => fetchAPI('/candidate/home'),
+    getAssessments: async () => fetchAPI('/candidate/assessments'),
+    getProfile: async () => fetchAPI('/candidate/profile')
+};
